feat: Implement EfilingPaymentResponseBody component for payment response handling
