+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/pages/citizen/FileCase/EFilingPaymentResponse.js
@@ -1,23 +1,6 @@
-import { Banner, Card, CardLabel, CardText, CloseSvg, Modal, TextArea } from "@egovernments/digit-ui-react-components";
-import React, { useMemo, useState } from "react";
-import Button from "../../../components/Button";
+import React from "react";
 import { useHistory, useLocation } from "react-router-dom/cjs/react-router-dom.min";
-import CustomCopyTextDiv from "../../../components/CustomCopyTextDiv";
-import SelectCustomNote from "../../../components/SelectCustomNote";
-import { Urls } from "../../../hooks";
-import useDownloadCasePdf from "../../../hooks/dristi/useDownloadCasePdf";
-
-const customNoteConfig = {
-  populators: {
-    inputs: [
-      {
-        infoHeader: "CS_COMMON_NOTE",
-        infoText: "PAYMENT_FAILED_NOTE_MSG",
-        infoTooltipMessage: "CS_NOTE_TOOLTIP_CASE_TYPE",
-      },
-    ],
-  },
-};
+import EfilingPaymentResponseBody from "../../../components/shared/EfilingPaymentResponseBody";
 
 const mockSubmitModalInfo = {
   header: "CS_PAYMENT_SUCCESSFUL",
@@ -27,10 +10,6 @@ const mockSubmitModalInfo = {
   isArrow: false,
   showTable: true,
 };
-
-
-
-
 
 function EFilingPaymentResponse({ t, setShowModal, header, subHeader, submitModalInfo = mockSubmitModalInfo, amount = 2000, path }) {
   const history = useHistory();
@@ -44,87 +23,25 @@ function EFilingPaymentResponse({ t, setShowModal, header, subHeader, submitModa
 
   const triggerSurveyContext = receiptData?.casePrevStatus === "PENDING_PAYMENT" ? "FILING_PAYMENT" : "DEFECT_CORRECTION_PAYMENT";
 
-  const commonProps = {
-    whichSvg: "tick",
-    headerStyles: { fontSize: "32px" },
-    style: { minWidth: "100%", marginTop: "10px" },
-  };
-
-  const bannerProps = isSuccess
-    ? {
-      ...commonProps,
-      successful: true,
-      message: t(submitModalInfo?.header),
-    }
-    : {
-      ...commonProps,
-      successful: false,
-      message: t("CS_PAYMENT_FAILED"),
-    };
-  const { downloadPdf } = useDownloadCasePdf();
   return (
     <div className=" user-registration">
-      <div className="e-filing-payment" style={{ minHeight: "100%", height: "100%" }}>
-        <Banner {...bannerProps} />
-        {submitModalInfo?.subHeader && isSuccess && <CardLabel className={"e-filing-card-label"}>{t(submitModalInfo?.subHeader)}</CardLabel>}
-        {receiptData ? (
-          <CustomCopyTextDiv
-            t={t}
-            keyStyle={{ margin: "8px 0px" }}
-            valueStyle={{ margin: "8px 0px", fontWeight: 700 }}
-            data={receiptData?.caseInfo}
-            tableDataClassName={"e-filing-table-data-style"}
-            tableValueClassName={"e-filing-table-value-style"}
-          />
-        ) : (
-          <SelectCustomNote t={t} config={customNoteConfig} />
-        )}
-        <div className="button-field" style={{ width: "100%", marginTop: 16 }}>
-          {!fileStoreId && caseId ? (
-            <Button
-              variation={"secondary"}
-              className={"secondary-button-selector"}
-              label={t("Retry Payment")}
-              labelClassName={"secondary-label-selector"}
-              onButtonClick={() => {
-                triggerSurvey(triggerSurveyContext, () => {
-                  history.push(`${path}/e-filing-payment?caseId=${caseId}`);
-                });
-              }}
-            />
-          ) : (
-            <Button
-              style={{
-                display: "flex",
-                color: "#505A5F",
-                textDecoration: "none",
-                // width: 250,
-                whiteSpace: "nowrap",
-                overflow: "hidden",
-                textOverflow: "ellipsis",
-              }}
-              variation={"secondary"}
-              className={"secondary-button-selector"}
-              label={t("CS_PRINT_RECEIPT")}
-              labelClassName={"secondary-label-selector"}
-              onButtonClick={() => {
-                downloadPdf(tenantId, fileStoreId);
-              }}
-            />
-          )}
-
-          <Button
-            className={"tertiary-button-selector"}
-            label={t("CS_GO_TO_HOME")}
-            labelClassName={"tertiary-label-selector"}
-            onButtonClick={() => {
-              triggerSurvey(triggerSurveyContext, () => {
-                history.push(`/${window?.contextPath}/citizen/dristi/home`);
-              });
-            }}
-          />
-        </div>
-      </div>
+      <EfilingPaymentResponseBody
+        isSuccess={isSuccess}
+        receiptData={receiptData}
+        fileStoreId={fileStoreId}
+        caseId={caseId}
+        submitModalInfo={submitModalInfo}
+        onRetry={() => {
+          triggerSurvey(triggerSurveyContext, () => {
+            history.push(`${path}/e-filing-payment?caseId=${caseId}`);
+          });
+        }}
+        onGoHome={() => {
+          triggerSurvey(triggerSurveyContext, () => {
+            history.push(`/${window?.contextPath}/citizen/dristi/home`);
+          });
+        }}
+      />
       {SurveyUI}
     </div>
   );
